chore: limit concurrent video controller init (#1886)
## Description
All changes to fix PlatfromException on android which are especially
good reproducible after pull to refresh. I've been sending the build to
QA before sending for review to make sure the fix is proper;
2 main changes here:
1. added ConcurrencyGate to make sure only specified amount of parallel
video player inits are in action. And set it to 1 for android;
2. added cancelInit which tracks if initialisation should be canceled in
flight (scroll off / dispose / refresh) and clean up resources properly
in that case;

## Type of Change
- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Refactoring
- [ ] Documentation
- [x] Chore