feat: improved media fallback (#1888)
## Description
This PR improves the fallback for images and videos.
If media fails to be loaded, we're taking the author relays and trying
to download the media from those.
If it fails, we're trying to update the author relays from identity and
try with updated relays (if any).

## Additional Notes
Also fixes infinity loop in case of image download error.

## Task ID
ION-2454

## Type of Change
- [x] Bug fix
- [x] New feature
- [ ] Breaking change
- [ ] Refactoring
- [ ] Documentation
- [ ] Chore