fix(onboarding): do not update attestation if it is already done for the current device (#1913)
## Description
This PR fixes an issue when the app adds the same, already delegated,
device pubkey on onboarding.

Details:
User creation involves 2 steps:
1. 10100 + 10002
2. 0, 3, 30015, ....

And if the app sends the first bundle of events successfully but fails
with the second one, the onboarding is considered as not completed. Next
time user retires, the app builds 10100 again by adding the current
device pubkey to the existing 10100 that leads to duplicate pubkeys
there. Fix is to check if the delegation is done for the current pubkey
and skip the 10100 assembling / sending in this case.

## Additional Notes
N/A

## Task ID
ION-3711

## Type of Change
- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Refactoring
- [ ] Documentation
- [ ] Chore