fix: filter counter results by event id (#1928)
## Description
This PR fixes

1. Incorrect counters REQ filter for 5400 results - 6400/7000.
Previously, we requested all 6400 and 7000 events for the current user
pubkey while waiting for the result of 5400. This caused the backend to
return all stored 6400 and 7000 events for the given user.
This PR changes the behavior to request only the 6400 and 7000 events
related to the specific 5400 event.

2. Redundant requesting of stories counters.
Previously, we requested total number of stories in a loop for each
story. The PR changes it to request it only once for each uniq story
author.

## Task ID
ION-3719

## Type of Change
- [X] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Refactoring
- [ ] Documentation
- [ ] Chore