fix: stories feed list UI (#1983)
## Description
This PR fixes several feed stories issues:
1. When a story is added to the list, but no metadata for that user is
loaded to the cache yet, we have a UI issues that `ListView.builder`
already have the "empty" item with paddings rendered.
Solution: add paddings inside the item so they are applied only when
user metadata is available.
2. Since `IonConnectAvatar` uses async `userMetadataProvider` that
returns `AsyncValue`, we always have `loading` state even tho metadata
is available right away from the cache.
Solution: add `sync` version of `userMetadataProvider`.
3. The `story_viewer_page` has logic to start the story player from the
first unseen story. This was broken because as soon as a story was
opened, it was marked as "seen." As a result, `viewedStories` was
updated, the hook recalculated a new initialStoryIndex, and the viewer
immediately jumped to the next story. This repeated until the last story
was opened.
Solution: run it once.
4. User stories are resorted when a new item is added to the stories
list.
Solution: rework the `viewedStoriesProvider` to keep the persistent list
of seen events' references and use it in all the places to avoid
recalculations / rebuilds.
5. Using `getSingleFile` instead of `downloadFile` for stories preload -
it doesn't download assets if they are already in cache.
6. Wait for N stories to be loaded before hiding the placeholder.

## Task ID
ION-2622
ION-3780

## Type of Change
- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Refactoring
- [ ] Documentation
- [ ] Chore

## Screenshots
before:


https://github.com/user-attachments/assets/838c59a6-46ed-4ae4-9ad8-5d56cb758d4d

after:


https://github.com/user-attachments/assets/49fa4e71-08ae-4f5f-96fd-94029854431c