feat(wallet): replace in-memory cache with DB for NFTs (#1993)
## Description
This PR replaces NFTs in-memory cache with DB table. Also it saves NFT
metadata into the table.

Known issue: `currentNftsProvider` might wait for a lot of futures if
user has a lot of NFTs. I plan to fix it in the next PR, where I'm gonna
implement pagination for loading NFTs.

## Type of Change
- [x] New feature