fix: make sure there is only one request filter (#2220)
## Description
With more than 1 request filter in pagination requests there can be a
situation
when we get created at of the last event in the stream but it is correct
only for one of the filters.
So here I make sure we always has only one filter for pagination
requests to fix such scenarios.

## Type of Change
- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Refactoring
- [ ] Documentation
- [ ] Chore