chore: rework fetch feed flow  (#2303)
## Description
This PR changes several aspects of fetching feed:
1. We fetch everything concurrently now: following + global + interests
data sources, explore + top + trending modifiers;
2. We fetch each page on the feed in 2 steps - first one with overflow
of the "interests" source because in most cases "following" and "global"
sources don't have enough data to show, then with the second step we
fetch the remaining number of events to fill the viewport;
3. We cache ranked relevant user relays to reduce the initial feed fetch
delay when determining which relays to query for data;
4. We change the `mostUsers` to `bestLatency` strategy when query
multiple relays.

## Additional Info
Fix the case where the cached config model is outdated and the app fails
to parse it — for example, when a new required field has been added to
the config model and the saved version doesn’t include it. In such
cases, we should re-fetch the config with `version=0`, otherwise the
backend returns `204`, and we won’t receive the updated values.

## Task ID
ION-4148

## Type of Change
- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Refactoring
- [ ] Documentation
- [x] Chore

## Videos
Before:


https://github.com/user-attachments/assets/64550786-00ea-4e38-a3aa-d400428a1565

After:


https://github.com/user-attachments/assets/2aa452e6-15ae-457d-be88-306a17a9c3ae