fix: precache picture in a separate zone to handle network exceptions (#2318)
## Description
We have a huge amount of network errors in sentry during precaching
pictures, They are propogated to global error handlers because they are
not awaited and handled inside both ImageCacheManager and
SvgNetworkLoader. So I put it in a separate zone to not this errors go
to global error handlers and to sentry.
Get the right cache manager for wallet coin icons and add a placeholder
icon for SVGs

## Type of Change
- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Refactoring
- [ ] Documentation
- [ ] Chore