feat: handle kind21750 -> kind3 on followers list (#2340)
## Description
This PR handles `EventsMetadataEntity->FollowListEntity` event that we
receive when fetching followers.

`FollowListEntity` are quite heavy and we don't need those data in this
case, so instead of original `FollowListEntity` events we receive
`EventsMetadataEntity->FollowListEntity` events with no data. We take
`.createdAt` for pagination and `.masterPubkey` from those wrapped
`FollowListEntity` events to build the list.

## Task ID
ION-4200

## Type of Change
- [ ] Bug fix
- [x] New feature
- [ ] Breaking change
- [ ] Refactoring
- [ ] Documentation
- [ ] Chore