feat: watch user avatar once to avoid refetches (#2353)
## Description
This PR handles the case when we first get user metadata from identity
and it has the original avatar url and then we fetch user metadata from
relays and there we're using the thumbnail url. To avoid downloading
both images, we're just using the first one to arrive.

## Task ID
N/A

## Type of Change
- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Refactoring
- [ ] Documentation
- [ ] Chore