feat: android soft update (#2397)
feat: implement in-app update feature for Android

## Description
This commit introduces an in-app update feature for Android, allowing
users to check for and initiate updates directly within the app. It
includes a new service for managing update checks and installation
status, as well as UI components for displaying update prompts.

## Task ID
ION-4158

## Type of Change
- [ ] Bug fix
- [x] New feature
- [ ] Breaking change
- [ ] Refactoring
- [ ] Documentation
- [x] Chore