fix: banuba unsupported size error during export (#2616)
## Description
Do not set VideoResolution.Original in custom ExportParams.

## Additional Notes
Root cause is that some devices does not support high resolution videos
processing, and forcing to use original resolution makes app to fail
with an exception. Fix is about applying VideoResolution.Auto which,
according to Banuba Support, "exports in the best possible quality that
are integrated in the VIdeo Editor SDK".
But I noticed that on my Sumsung S23 it uses FHD even if it supports 4K
videos. This fix is still acceptable cause we have a requirement to
export video at least in FHD

## Task ID
ION-4523

## Type of Change
- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Refactoring
- [ ] Documentation
- [ ] Chore