fix: camera stream crash on iphone17 (#2654)
## Description
Camera stream crashing on iPhone 17 with `Unsupported pixel format type`
when using `ResolutionPreset.max`

## Additional Notes
As this new btp2 format is not supported yet by camera plugin the only
option for now is to not use highest resolution on iPhone17.
 I have prepared 2 different fixes for it:
1. try to init camera with the highest resolution and if it fails -
fallback to the next highest resolution option. I would say that it is
the preferred option, because it is device model agnostic and should
work for all devices. The thing is that I cannot check it on real device
so I cannot be sure that iOS native part still do not crash when we try.
2. just uses ResolutionPreset.ultraHigh instead of ResolutionPreset.max
for iPhone 17.

This is the 1 option implementation, agreed with PO to merged it.

## Task ID
ION-4524

## Type of Change
- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Refactoring
- [ ] Documentation
- [ ] Chore

(cherry picked from commit e94b7a776b39bb9b894e5db352665a7f375d03e9)