fix: several markdown issues (check description) (#2690)
Fixes:
Did add a post with a link on it from this latest build, and on old
builds from store link is not visible
Feed. Bold font is not displaying
Mentioned used is displayed as regular text on published post
Feed. “Copy” menu item is not available for post text after publishing

<!-- Add any extra context or relevant information here. -->

4657
4650
4656
4655

- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Refactoring
- [ ] Documentation
- [ ] Chore

<!-- Include screenshots to demonstrate any UI changes. -->
<!-- <img width="180" alt="image" src="image_url_here"> -->