fix: iphone17 camera crash (#2712)
## Description
Camera stream crashing on iPhone 17 with Unsupported pixel format type
when using ResolutionPreset.max

## Additional Notes
As this new btp2 format is not supported yet by camera plugin the only
option for now is to not use highest resolution on iPhone17.
As a workaround we use ResolutionPreset.ultraHigh instead of
ResolutionPreset.max for iPhone 17.

Includes revert of the previous fix tryout. [PR
2654](https://github.com/ice-blockchain/ion-framework/pull/2654)

## Task ID
ION-4524

## Type of Change
- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Refactoring
- [ ] Documentation
- [ ] Chore