feat(tokenized communities): show RequiredBscWalletDialog (#2757)
## Description
This PR added a `RequiredBscWalletDialog`. We show on app start for
logged in user in case they don't have a BSC wallet. The dialog creates
the wallet on "Get Started" pressed. The action is mandatory and can not
be dismissed.

## Task ID
ION-4648

## Type of Change
- [x] New feature

## Screenshots
<img width="350" alt="image"
src="https://github.com/user-attachments/assets/340c1927-c67c-4c25-927c-9900178ffadf"
/>