chore: set user metadata cache expiration from env (#2962)
## Description
This PR changes the user metadata caching to 2 minutes because we need
to refetch those from time to time to get the fresh dependencies
(31175);

## Additional Notes
Also this PR adds `DependencyEventsHandler` that just unwraps the 21750
events and caches. Is needed because `kind[0/1/30175/30023]>kind31175`
returns the original event that `31175` points to and that event is
wrapped with 21750.

## Task ID
N/A

## Type of Change
- [ ] Bug fix
- [x] New feature
- [ ] Breaking change
- [x] Refactoring
- [ ] Documentation
- [x] Chore